requirements.txt
yaml
paho-mqtt
gevent